Support optional arguments
When processing optional arguments, there's three possible behaviors:
taking a single value (in the case of an optional single-valued type),
taking multiple values (in the case of an optional Array-like type), or
taking no values (in the case of not having enough required positional
arguments to take).

We take the greedy strategy here: as long as we have enough remaining
items to potentially fulfill our promises about required arguments on
the command line, we take it.

Note that this greedy strategy doesn't work well when dealing with mixed
positional arguments and options. However, heuristics at config parse
time for detecting potentially ambiguous grammars can handle that.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>